Refactor: split style_bert_vits2.nlp.english package
